Started fixing for loading multiple skins
